Unify sripts
